+++ wgpu-core/src/resource.rs
@@ -2613,6 +2613,10 @@ impl TextureView {
             label: desc.label.to_string(),
         })
     }
+
+    pub fn texture(&self) -> &Arc<Texture> {
+        &self.parent
+    }
 }
 
 #[derive(Clone, Debug, Error)]